+++ src/lib/State/Processor.ts
@@ -6,14 +6,14 @@ export type StateProcessorInput<T> = T | Array<T> | null;
 export interface StateProcessor<T> {
 	process(
 		data: StateProcessorInput<T>,
-		operation: CrudOperation<T>,
+		operation: CrudOperation,
 		requestParameters: KeyValueObject
 	): void;
 }
 
 export type StateProcessorCallback<T> = (
 	data: StateProcessorInput<T>,
-	operation: CrudOperation<T>,
+	operation: CrudOperation,
 	requestParameters: KeyValueObject
 ) => void;
 
@@ -26,7 +26,7 @@ export class CallbackStateProcessor<T> implements StateProcessor<T> {
 
 	process(
 		data: StateProcessorInput<T>,
-		operation: CrudOperation<T>,
+		operation: CrudOperation,
 		requestParameters: KeyValueObject
 	): void {
 		return this._callback(data, operation, requestParameters);